trying to make the code more elegant
One way to make a loop more elegant is by reducing the number of if
statements. The previous version used a while(true) : (i -= 1) and then had an if
statement at the bottom to check if i == 0. The new version checks if i
< 0, but decrements i at the top of the loop. This seems to be more
elegant because while loops have this awkward pattern when traversing
from the end of a slice to the front.