heapify rewrite day 1.
I'll admit that I used inspiration from the linux kernel lib/sort.c

The key insight is that instead of having to do 2 compares in each
branch when shifting down, (finding the max value between the parent,
and both children), reduce the number of compares and swaps with the
following:
Since the invariant of the heap is that the parent is larger than
either of its two children, we'll build up from a simple example.

   a
  / \
 b   c

For a single binary tree, we can heapify this with a 2 compares
and a possible swap.
1. m <- if (b > c) b else c;
2. if (a < m) swap(a, m);

Now if in the above example, b and c are trees themselves, but those
trees are heapified, then we make a small adjustment

1. Descend the tree at a head node by following the larger child until
   you reach a leaf node.
2. Ascend the tree until you find a node (cursor) that smaller than head (a)
   if you don't find any, then the heap invariant is satisifed, break;
3. Since all the children of (cursor) node has the heap invariant, we
   can essentially rotate these values until the heap invariant is
   satisfied.

   If a >= b >= c >= d, but you have a tree path of d->a->b->c then you can sort
   this by doing essentially a bubble sort. With only the index or
   address of b, we can swap the cursor with the head and then move the
   cursor to its parent.

   swap(d,c)=> c->a->b->d
   move cursor to b, head is pointing at c
   swap(b,c)=> b->a->c->d
   move cursor to a, head is pointing at b
   swap(a,b)=> a->b->c->d
   move cursor to a, cursor is at head, break;

So instead of doing 2 compares at every branch for depth N, we do N
compares to descend into the tree and M (M < N) compares for ascending we find
a slot for the head for M swaps.